break down block.js into separate parts
