+++ src/blocks/expand/styles/style.scss
@@ -4,7 +4,7 @@
  * CSS for both Frontend+Backend.
  */
 
-.ahg-card {
+ .ahg-card {
 	border-radius: 4px;
 	margin: 30px 0;
 	overflow: hidden;
